Add max connections support
It would be quite helpful, especially for datasets hosted on MongoDB Atlas and similar services that limit connections, to be able to set the maximum amount of connections Josh can concurrently use. This pull request adds that feature, and allows users to provide `maxConnections` in the provider options object to set it.